jobboss-odoo-connector/src/jobboss/client.py
import psycopg2

class JobbossClient:
    def __init__(self, db_url, username, password):
        self.db_url = db_url
        self.username = username
        self.password = password

    def connect(self):
        # Conexión a la base de datos
        return psycopg2.connect(self.db_url, user=self.username, password=self.password)

    def create_record(self, data):
        # Implementar la lógica para crear un registro en Jobboss
        pass

    def read_record(self, record_id):
        # Implementar la lógica para leer un registro de Jobboss
        try:
            connection = self.connect()
            cursor = connection.cursor()
            query = "SELECT * FROM your_table_name WHERE id = %s"
            cursor.execute(query, (record_id,))
            record = cursor.fetchone()  # Obtiene el primer registro que coincide
            cursor.close()
            connection.close()
            return record
        except psycopg2.Error as e:
            print(f"Error al leer el registro: {e}")
            return None

    def update_record(self, record_id, data):
        # Implementar la lógica para actualizar un registro en Jobboss
        pass

    def delete_record(self, record_id):
        # Implementar la lógica para eliminar un registro de Jobboss
        pass