jobboss-odoo-connector/requirements.txt
fastapi
httpx
pydantic
sqlalchemy
uvicorn
pyyaml